requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-celery-beat==2.4.0
django-config==0.1.4
django-filter==22.1
django-modeltranslation==0.18.9
django-rosetta==0.9.8
django-timezone-field==5.0
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.4.0
polib==1.1.1
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
